Move private methods to its own module.
